+++ module/party.js
@@ -2,7 +2,7 @@ import { WwnPartySheet } from "./dialog/party-sheet.js";
 
 export const addControl = (object, html) => {
   let control = `<button class='wwn-party-sheet' type="button" title='${game.i18n.localize('WWN.dialog.partysheet')}'><i class='fas fa-users'></i></button>`;
-  html.find(".fas.fa-search").replaceWith($(control))
+  html.find(".toggle-search-mode").before($(control));
   html.find('.wwn-party-sheet').click(ev => {
     showPartySheet(object);
   })